requirements.txt
wheel==0.23.0
Django>=1.6,<1.7
South==1.0
django-braces==1.4.0
django-model-utils==2.0.3
gevent==1.0.1
lxml==3.3.5
reppy==0.2.2
six==1.7.3
